+++ src/t8_schemes/t8_default/t8_default_tri/t8_dtri_bits.c
@@ -424,12 +424,12 @@ t8_dtri_compute_reference_coords (const t8_dtri_t *elem,
   const int           tet_orientation2 =
     (tet_orientation0 + ((type % 2 == 0) ? 2 : 1)) % 3;
 #endif
-  for (int coord = 0; coord < num_coords; ++coord) {
+  for (size_t coord = 0; coord < num_coords; ++coord) {
     /* offset defines, how many coordinates to skip in an iteration. */
 #ifndef T8_DTRI_TO_DTET
-    const int           offset = (2 + skip_coords) * coord;
+    const size_t        offset = (2 + skip_coords) * coord;
 #else
-    const int           offset = 3 * coord;
+    const size_t        offset = 3 * coord;
 #endif
     out_coords[offset + 0] = elem->x;
     out_coords[offset + 1] = elem->y;
